+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/RealMiniAppSpec.kt
@@ -37,15 +37,6 @@ class RealMiniAppSpec {
     }
 
     @Test
-    fun `should still be able to download miniapp for deprecated method`() =
-        runBlockingTest {
-            realMiniApp.create(miniAppInfo)
-            realMiniApp.create(miniAppInfo, miniAppMessageBridge)
-
-            verify(miniAppDownloader, times(2)).getMiniApp(TEST_MA_ID)
-        }
-
-    @Test
     fun `should invoke from MiniAppInfoFetcher when calling list miniapp`() = runBlockingTest {
         realMiniApp.listMiniApp()
 
@@ -65,11 +56,13 @@ class RealMiniAppSpec {
     @Test
     fun `should invoke from MiniAppDownloader and Displayer when calling create miniapp`() =
         runBlockingTest {
+            val getMiniAppResult = Pair(TEST_BASE_PATH, TEST_MA)
+            When calling miniAppDownloader.getMiniApp(TEST_MA_ID) itReturns getMiniAppResult
             realMiniApp.create(TEST_MA_ID, miniAppMessageBridge)
 
-            val basePath: String = verify(miniAppDownloader, times(1))
-                .getMiniApp(TEST_MA_ID)
-            verify(displayer, times(1)).createMiniAppDisplay(basePath, TEST_MA_ID, miniAppMessageBridge)
+            verify(miniAppDownloader, times(1)).getMiniApp(TEST_MA_ID)
+            verify(displayer, times(1))
+                .createMiniAppDisplay(getMiniAppResult.first, getMiniAppResult.second, miniAppMessageBridge)
         }
 
     @Test