+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/display/MiniAppWebviewTest.kt
@@ -2,6 +2,7 @@ package com.rakuten.tech.mobile.miniapp.display
 
 import android.content.Context
 import android.net.Uri
+import android.util.AndroidRuntimeException
 import android.view.ViewGroup
 import android.webkit.GeolocationPermissions
 import android.webkit.WebResourceRequest
@@ -14,9 +15,11 @@ import com.nhaarman.mockitokotlin2.atLeastOnce
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.times
 import com.nhaarman.mockitokotlin2.verify
+import com.rakuten.tech.mobile.miniapp.*
+import com.rakuten.tech.mobile.miniapp.TEST_BODY_CONTENT
 import com.rakuten.tech.mobile.miniapp.TEST_HA_NAME
-import com.rakuten.tech.mobile.miniapp.TEST_MA_ID
 import com.rakuten.tech.mobile.miniapp.TEST_URL_HTTPS_1
+import com.rakuten.tech.mobile.miniapp.TEST_URL_HTTPS_2
 import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
 import org.amshove.kluent.*
 import org.junit.Before
@@ -38,12 +41,12 @@ open class BaseWebViewTest {
     fun setup() {
         context = getApplicationContext()
         basePath = context.filesDir.path
-        webChromeClient = Mockito.spy(MiniAppWebChromeClient(context))
+        webChromeClient = Mockito.spy(MiniAppWebChromeClient(context, TEST_MA))
 
         miniAppWebView = MiniAppWebView(
             context,
             basePath = basePath,
-            appId = TEST_MA_ID,
+            miniAppInfo = TEST_MA,
             miniAppMessageBridge = miniAppMessageBridge,
             hostAppUserAgentInfo = TEST_HA_NAME,
             miniAppWebChromeClient = webChromeClient
@@ -57,12 +60,12 @@ class MiniAppWebviewTest : BaseWebViewTest() {
 
     @Test
     fun `for a given app id, creates corresponding view for the caller`() {
-        miniAppWebView.url shouldContain miniAppWebView.appId
+        miniAppWebView.url shouldContain miniAppWebView.miniAppInfo.id
     }
 
     @Test
     fun `should have corrected load url format`() {
-        miniAppWebView.getLoadUrl() shouldEqual "https://mscheme.${miniAppWebView.appId}/miniapp/index.html"
+        miniAppWebView.getLoadUrl() shouldEqual "https://mscheme.${miniAppWebView.miniAppInfo.id}/miniapp/index.html"
     }
 
     @Test
@@ -102,7 +105,7 @@ class MiniAppWebviewTest : BaseWebViewTest() {
         miniAppWebView = MiniAppWebView(
             context,
             basePath = basePath,
-            appId = TEST_MA_ID,
+            miniAppInfo = TEST_MA,
             miniAppMessageBridge = miniAppMessageBridge,
             hostAppUserAgentInfo = "",
             miniAppWebChromeClient = webChromeClient
@@ -123,9 +126,9 @@ class MiniAppWebviewTest : BaseWebViewTest() {
     @Test
     fun `each mini app should have different domain`() {
         val miniAppWebViewForMiniapp1 = MiniAppWebView(
-            context, miniAppWebView.basePath, "app-id-1", miniAppMessageBridge, TEST_HA_NAME)
+            context, miniAppWebView.basePath, TEST_MA, miniAppMessageBridge, TEST_HA_NAME)
         val miniAppWebViewForMiniapp2 = MiniAppWebView(
-            context, miniAppWebView.basePath, "app-id-2", miniAppMessageBridge, TEST_HA_NAME)
+            context, miniAppWebView.basePath, TEST_MA.copy(id = "app-id-2"), miniAppMessageBridge, TEST_HA_NAME)
         miniAppWebViewForMiniapp1.url shouldNotBeEqualTo miniAppWebViewForMiniapp2.url
     }
 
@@ -147,7 +150,7 @@ class MiniAppWebClientTest : BaseWebViewTest() {
     fun `should intercept request with WebViewAssetLoader`() {
         val webAssetLoader: WebViewAssetLoader =
             Mockito.spy((miniAppWebView.webViewClient as MiniAppWebViewClient).loader)
-        val webViewClient = MiniAppWebViewClient(webAssetLoader,
+        val webViewClient = MiniAppWebViewClient(context, webAssetLoader,
             "custom_domain", "custom_scheme")
         val webResourceRequest = getWebResReq(TEST_URL_HTTPS_1.toUri())
 
@@ -160,24 +163,43 @@ class MiniAppWebClientTest : BaseWebViewTest() {
     @Test
     fun `should redirect to custom domain when only loading with custom scheme`() {
         val webAssetLoader: WebViewAssetLoader = (miniAppWebView.webViewClient as MiniAppWebViewClient).loader
-        val customDomain = "https://mscheme.${miniAppWebView.appId}/"
-        val webViewClient = Mockito.spy(MiniAppWebViewClient(webAssetLoader, customDomain,
-            "mscheme.${miniAppWebView.appId}://"))
+        val customDomain = "https://mscheme.${miniAppWebView.miniAppInfo.id}/"
+        val webViewClient = Mockito.spy(MiniAppWebViewClient(context, webAssetLoader, customDomain,
+            "mscheme.${miniAppWebView.miniAppInfo.id}://"))
 
         val displayer = Mockito.spy(miniAppWebView)
 
         webViewClient.onReceivedError(displayer, webResourceRequest, mock())
         webViewClient.onReceivedError(displayer,
-            getWebResReq("mscheme.${miniAppWebView.appId}://".toUri()), mock())
+            getWebResReq("mscheme.${miniAppWebView.miniAppInfo.id}://".toUri()), mock())
 
         verify(webViewClient, times(1)).loadWithCustomDomain(displayer, customDomain)
+    }
+
+    @Test
+    fun `should open phone dialer when there is telephone scheme`() {
+        val webAssetLoader: WebViewAssetLoader = (miniAppWebView.webViewClient as MiniAppWebViewClient).loader
+        val webViewClient = Mockito.spy(MiniAppWebViewClient(context, webAssetLoader, "",
+            "mscheme.${miniAppWebView.miniAppInfo.id}://"))
+        val displayer = Mockito.spy(miniAppWebView)
+        val phoneUri = "tel:123456"
+
+        @Suppress("SwallowedException")
+        try {
+            webViewClient.shouldOverrideUrlLoading(displayer, webResourceRequest)
+            webViewClient.shouldOverrideUrlLoading(displayer, getWebResReq(phoneUri.toUri()))
+        } catch (e: AndroidRuntimeException) {
+            // context here is not activity
+        }
+
+        verify(webViewClient, times(1)).openPhoneDialer(phoneUri)
     }
 
     @Test
     fun `should not intercept mime type for regular cases`() {
         val webResourceResponse = WebResourceResponse("", "utf-8", ByteArrayInputStream("".toByteArray()))
         val webClient = miniAppWebView.webViewClient as MiniAppWebViewClient
-        val request = getWebResReq("mscheme.${miniAppWebView.appId}://".toUri())
+        val request = getWebResReq("mscheme.${miniAppWebView.miniAppInfo.id}://".toUri())
         val secondRequest = Mockito.spy(getWebResReq("test.js".toUri()))
 
         webClient.interceptMimeType(webResourceResponse, request)
@@ -218,7 +240,7 @@ class MiniAppWebChromeTest : BaseWebViewTest() {
 
     @Test
     fun `bridge js should be null when js asset is inaccessible`() {
-        val webClient = MiniAppWebChromeClient(mock())
+        val webClient = MiniAppWebChromeClient(mock(), mock())
         webClient.bridgeJs shouldBe null
     }
 
@@ -235,6 +257,18 @@ class MiniAppWebChromeTest : BaseWebViewTest() {
         webChromeClient.onGeolocationPermissionsShowPrompt(null, null)
 
         verify(geoLocationCallback, times(1)).invoke("", true, false)
+    }
+
+    @Test
+    fun `should override js dialog event`() {
+        webChromeClient.onJsAlert(
+            miniAppWebView, TEST_URL_HTTPS_2, TEST_BODY_CONTENT, mock()) shouldBe true
+        webChromeClient.onJsConfirm(
+            miniAppWebView, TEST_URL_HTTPS_2, TEST_BODY_CONTENT, mock()) shouldBe true
+        webChromeClient.onJsPrompt(
+            miniAppWebView, TEST_URL_HTTPS_2, TEST_BODY_CONTENT, TEST_VALUE, mock()) shouldBe true
+        webChromeClient.onJsPrompt(
+            miniAppWebView, TEST_URL_HTTPS_2, TEST_BODY_CONTENT, null, mock()) shouldBe true
     }
 }