+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/settings/SettingsMenuActivity.kt
@@ -1,5 +1,7 @@
 package com.rakuten.tech.mobile.testapp.ui.settings
 
+import android.app.Activity
+import android.content.Intent
 import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
@@ -7,6 +9,7 @@ import android.view.Gravity
 import android.view.Menu
 import android.view.MenuItem
 import android.widget.Toast
+import androidx.appcompat.widget.AppCompatEditText
 import androidx.appcompat.widget.Toolbar
 import com.rakuten.tech.mobile.miniapp.MiniApp
 import com.rakuten.tech.mobile.miniapp.MiniAppSdkException
@@ -40,7 +43,7 @@ class SettingsMenuActivity : BaseActivity() {
         override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {}
 
         override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {
-            enableSaveView()
+            validateInputIDs()
         }
     }
 
@@ -91,7 +94,7 @@ class SettingsMenuActivity : BaseActivity() {
     }
 
     private fun renderAppSettingsScreen() {
-        textInfo.text = createSettingsInfo()
+        textInfo.text = createBuildInfo()
         editAppId.setText(settings.appId)
         editSubscriptionKey.setText(settings.subscriptionKey)
         switchTestMode.isChecked = settings.isTestMode
@@ -99,17 +102,33 @@ class SettingsMenuActivity : BaseActivity() {
         editAppId.addTextChangedListener(settingsTextWatcher)
         editSubscriptionKey.addTextChangedListener(settingsTextWatcher)
 
-        enableSaveView()
+        validateInputIDs()
     }
 
-    private fun createSettingsInfo(): String {
-        return "Build " + getString(R.string.miniapp_sdk_version) + " - " +
-                getString(R.string.build_version)
+    private fun createBuildInfo(): String {
+        val sdkVersion = getString(R.string.miniapp_sdk_version)
+        val buildVersion = getString(R.string.build_version)
+        return "Build $sdkVersion - $buildVersion"
     }
 
-    private fun enableSaveView() {
-        saveViewEnabled =
-            !(editAppId.text.toString().isInvalidUuid() || editSubscriptionKey.text.isNullOrEmpty())
+    internal fun validateInputIDs() {
+        val isAppIdInvalid = editAppId.text.toString().isInvalidUuid()
+
+        saveViewEnabled = !(isInputEmpty(editAppId)
+                || isInputEmpty(editSubscriptionKey)
+                || isAppIdInvalid)
+
+        if (isInputEmpty(editAppId) || isAppIdInvalid) {
+            editAppId.error = getString(R.string.error_invalid_input)
+        }
+
+        if (isInputEmpty(editSubscriptionKey)) {
+            editSubscriptionKey.error = getString(R.string.error_invalid_input)
+        }
+    }
+
+    private fun isInputEmpty(input: AppCompatEditText): Boolean {
+        return input.text.toString().isEmpty() || input.text.toString().isBlank()
     }
 
     private fun updateSettings(appId: String, subscriptionKey: String, isTestMode: Boolean) {
@@ -136,7 +155,7 @@ class SettingsMenuActivity : BaseActivity() {
                     settingsProgressDialog.cancel()
                     val toast =
                         Toast.makeText(this@SettingsMenuActivity, error.message, Toast.LENGTH_LONG)
-                    toast.setGravity(Gravity.TOP, 0, 0)
+                    toast.setGravity(Gravity.BOTTOM, 0, 100)
                     toast.show()
                 }
             }
@@ -146,7 +165,12 @@ class SettingsMenuActivity : BaseActivity() {
     private fun navigateToPreviousScreen() {
         when (intent.extras?.getString(MENU_SCREEN_NAME)) {
             MINI_APP_LIST_ACTIVITY -> {
-                raceExecutor.run { launchActivity<MiniAppListActivity>() }
+                val intent = Intent(this, MiniAppListActivity::class.java).apply {
+                    addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_NEW_TASK)
+                }
+                setResult(Activity.RESULT_CANCELED, intent)
+                startActivity(intent)
+                finish()
             }
             MINI_APP_INPUT_ACTIVITY -> {
                 raceExecutor.run { launchActivity<MiniAppInputActivity>() }